fix(media): route all uploads through a shared raw-binary helper
The block editor's image/file pickers and the SEO image field sent uploads as
multipart/form-data (FormData). The server's handleUpload reads the request body
via request.arrayBuffer() and never parses multipart, so those uploads failed
while the media library (already on the raw-binary protocol) worked.

Add uploadMedia() to media-fetch.ts — the single upload path: raw file body,
real MIME as Content-Type, filename in x-cms-filename. Migrate block-form.ts,
page-editor.ts, and media.ts to it, removing the divergent copies that let two
call sites fall behind when the protocol changed.
