+++ routes/admin/client/page-editor.ts
@@ -24,6 +24,7 @@ import {
 } from './common.js';
 import { ct } from '../i18n/client.js';
 import { mountBlockForm, type BlockFormHandle, type ArrayLimitInfo } from './block-form.js';
+import { uploadMedia } from './media-fetch.js';
 
 const trashIconSvg =
   '<svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M3 6h18"/><path d="M19 6v14c0 1-1 2-2 2H7c-1 0-2-1-2-2V6"/><path d="M8 6V4c0-1 1-2 2-2h4c1 0 2 1 2 2v2"/><line x1="10" y1="11" x2="10" y2="17"/><line x1="14" y1="11" x2="14" y2="17"/></svg>';
@@ -737,15 +738,9 @@ export function initPageEditor(): void {
   }
 
   async function uploadSeoImage(file: File): Promise<void> {
-    const formData = new FormData();
-    formData.append('file', file);
-
-    const data = await fetchJson<{ url?: string }>('/cms/api/upload', {
-      method: 'POST',
-      headers: authHeaders(false),
-      body: formData,
-    });
-
+    const response = await uploadMedia(file);
+    const data = await response.json().catch(() => ({})) as { url?: string; error?: string };
+    if (!response.ok) throw new Error(data.error || 'Request failed');
     if (data.url) updateSeoImagePreview(data.url);
   }
